Add temporary "--problems ABC,DEF" to resolver
Add option to filter by problems for the PacNW contest. This argument can only be used when you use the --groups option, and it filters to problems in the given list, i.e. group 1 would only show problems A,B, and C, and group 2 would only show problems D, E, and F.