Clients only try websocket when connecting to a CDS
This is the other half of a recent change; clients can tell if the remote server is a CDS based on the HTTP header, and won't try to connect the admin websocket if connecting to another contest source (e.g. CCS). FWIW there will be a temporary 'regression' if you use newer clients with an older CDS then they won't try to connect.

Sorry about the other formatting on versioning - Eclipse insisted on a few of them, then I made a few myself.