Remove --problems and allow resolver to run multiple contests in order
Adds the ability to specify two or more contests through the resolver command line by using "contestAurl&contestBurl user password". When more than one contest is specified the resolver will resolve one, then return to the splash screen and resolve the second (or third, etc). Undocumented for now until we decide on the right commandline syntax.
Also removes the --problems argument, since that can now be duplicated by using the CDS' view filtering along with the resolver multi-contest support, which is a much cleaner and more general purpose way of supporting this.