Delete client cache before reopening, avoid reconnection NPE
Issue 87. At PacNW, some clients would detect that the feed had changed, disconnect, and then reconnect using the same cache and enter a loop. I cannot reproduce, but one possibility is that on Windows the FileOutputStream behaviour is different and the cache is not thrown out because the file has already been reopened. Closing this loophole and adding another log message in case this wasn't the problem.