Automatic reloading of cdsConfig.xml
Watches cdsConfig.xml and automatically reloads as necessary every 4s (because 5s felt too long...). Uses a hash of each contest's xml snippet (<contest> element) to tell if anything has changed. If not (even if moved or reordered in the file), the contest is not touched. If the contest is changed or removed it gets cleaned up and any new or modified contests get started.