Contest and Org getActualFormalName
When we implemented ITeam.getDisplayName() we added getActualDisplayName() so that the API was returning the correct/real value, but for an optional parameter clients could use one method with a fallback. The same issue existed for Contest and Organization formal names, so I fixed it here. Don't love the name, but can't think of anything better. Also fixed a few places where null checks were missing for getCountry().