Import from teams.json
Simplified TSVImporter so that it was just a straight file reader and cleaned up the exception handling. Moved the logic to DiskContestSource so that it looks for *.json, then *2.tsv, then .tsv. Technically we only need teams.json for now, but tested that orgs.json works fine too. Noticed that config validation was pretty poor and improved that, and found and fixed a few places where we weren't respecting team display name.