Remove support for XML feed everywhere but CDS
Removed support for all command line tools and connection dialogs for connecting to a socket + XML feed. Removed support for the submission file interface from the CDS but *left in the support for connecting to the XML socket and start time interface*. Removed references to the XML feed in the docs, generally tried to clean them up as I went, made the CDS 'not special' and just another Contest API source, and added DOMjudge to the list of verified Contest API sources.