Consistent command line parser
Created a single command line parser for all of the tools, which enforces a consistent pattern on the command line "contestURL/path [options]" where each option may have additional arguments. This forced some changes for consistency in a few of the tools, which I'll document in the issue. Corrected all the --help output to match.
Also updated the CDS to add a response header more consistently, and the tools (primarily pres admin) will confirm that you are actually connecting to the CDS instead of failing later.