Allow team rotation to be float
The CDS only allowed the team's desk rotation to be integers, but we made the spec support floating point values. Updating the contest model to support double, and getting rid of float/short in FloorMap utility and downstream effect on FloorGenerators and BalloonPrinter. Using this opportunity to check in the NAC floor map gen too.