Remove solution awards
The solution awards were 'non standard', didn't really mean anything, and always showed up as a difference when compared to DOMjudge. I've removed the award type and stopped generating them in both the CDS & award util. To avoid change in behaviour I did change the resolver so that if there are solution awards it'll show them, and if there aren't it'll show the number of problems solved for all teams that received a medal (which is all the solution award was anyway).