Automatic presentation switching
This change is a first pass at automatically switching presentations at a given contest time. It adds support in the admin UI for assigning a presentation at a given contest time, and the presentation server support for managing and firing these as events later on.
This change forgets the settings if the CDS has to be restarted. It waits for the first contest event after the given time (should have a real timer). It's simplistic. There's no way to see what you've set or clear it. However, I believe this is the minimum change set to try this out in sysops at the NAC and start to talk about whether it's a good thing or not, and if so how it should fit into the UI better.