Award Templates
Had a changeset going since Christmas, decided it was better to scale it back to MVP so that there's actually a chance to test and get feedback at NAC. This is the minimal change that adds support for assigning by award template to both the admin UI and CDS admin page. Minimal code change to existing code paths in order to minimize risk.