Contest location
The world balloon map was hardcoded to Porto. This provides support for a new contest location object in the Contest API (which will be proposed to CLICS later), and if set (via contest.json or CCS) the presentation will use it. Could be used for other map presentations later as well.