Admin add/remove objects
Commit contains three things: 1. Support for arbitrary add and delete of contest objects from the admin page. Changing objects here will change the contest model on the CDS and send out events, handy if, for instance, the CCS skips an event and you need to tweak the event feed instead of resetting everything. 2. A lock button on the countdown control section so that you can see the time but can't change it, requested to bring back by John at NAC. 3. Formatting of the admin page.