Balloon printout
At NAC and last year's WF we were asked to print a set of pages for the balloons - one big label per problem and then one page that shows all the balloons and their colours. This function has only been available in the ProblemSet util, which means we need to download and fire it up just for this.
This commit adds the identical function to the CDS: a new admin section with a link to a balloon printout with the same pages. To do this I moved the Balloon helper class and its images 'down' from presentations to the contest model - perhaps not ideal, but doesn't add any new dependencies and is <25K.