Resolver org logo side screen
Someone suggested another 'side-screen' for the resolver that would show all of the team logos on screen, and slowly fade each one out as they were resolved until you got to the winner. I actually wrote this on the way to NAC, but finally got it tidied up now. As logos are removed it automatically scales the remaining ones up to fit the screen.
This change required one 'internal' method on IOrganization so that the presentation can see exactly which logo sizes are available and know when to grab the next size up. This should probably be exposed differently (likely public interface for FileReferenceList?) but I thought it was ok for now.