Add back spare teams
As part of the last update to make floor maps part of the contest API I forgot to add back 'spare team desks' - basically an additional set of ITeams that have location data but nothing else, stored under the MapInfo so that they aren't real teams. The API support was there but the floor map didn't put the data in nor draw them if they were there.
This change adds the data to the MapInfo, consolidates drawing of teams between the two drawFloor() methods, and draws the spare teams on the map. I also rotated the team numbers on the video map, renamed 'balloon' to 'problem' in the floor-map.tsv (with backward compatibility), and fixed 2 minor issues in the 2016 and 2018 map generators. The next commit will remove the older generators completely, but I wanted to fix them first.