Simple error page for REST, hide admin page for non-admins
At NAC I tried to use one of the admin actions while logged in as balloon. To keep things simple, I'm just putting in a placeholder page when you're not an admin.
The pretty default error page was being returned for REST requests and was pretty ugly from command line or embedded as text, so I'm inserting a servlet in that can divert REST errors to the vanilla response text.