Add FPS parameter
Add a parameter to both the standalone and client presentations to show the frame rate onscreen and rename the variable/methods to be FPS consistently. We had this option many years ago and it was reduced to only being visible in the admin, but found a couple times where this was handier, e.g. testing a new presentation.