Website tweaks
Change the resolver font-awesome to 'trophy' to match the icon I made. Add a .project file to make it easier to edit in Eclipse (not that you should...). Change the WLP download to show Open Liberty vs WAR. Tweak to builds page to indent downloads and make download size column right-justified and smaller.