The Resolver Optimizer
I have been sitting on this since the flight home from last year's finals, time to check it in and not risk loss in a machine crash. :)
The 'resolver optimizer' is a command line tool that will look at a contest to determine all the possible resolution paths, narrow down to which are actually significant (e.g. teams with submissions to two or more problems, but not ones where they're both failed), shorten to which ones affect the top teams, determine which resolution is 'best', and then output a config file that will drive the resolver to use those resolution steps.
The delay in checking this in is both because nobody has been able to quantify what 'best' is, and because in most contests it doesn't seem to make much difference. So this is all still TBD, but I'd rather get this utility checked in for now and we can always adapt/remove later. IMHO this may turn out better as a util to spit out all the option points and give guidance, i.e. more data and easier for a human to decide which path they want vs automating the decision.