Reconnect unless contest is done updating
During two of the test contests at NAC the CDS received a normal (EOF-style) event feed termination from Kattis - i.e. no errors, the endpoint was just done giving data. The CDS interpreted this as 'ok, there's no more data to get' and didn't attempt to reconnect.
Whether this was some bizarre network glitch, a bug in Kattis, or someone restarting the feed incorrectly, it seems like the far safer approach is to attempt reconnect until a contest is done updating.
Moved the completion event outside of this too (don't say we're complete if we're going to reconnect). Bumped the reconnect time up to 5s, for no particular reason other than it seems more natural and IMHO when things go down it's typically for longer periods anyway.