Resolver API
Up until now the resolver has been a separate tool from everything else. A couple presentations have support for selecting a run or scrolling, but the resolver is a standalone tool that figures out the steps, launches its own UI, and drives everything.

This change moves the core of the resolver (the 'logic' that determines the steps and its utility classes) to the contest model project alongside the contest API. Two helper classes from the presentations (SubmissionInfo and SelectType) also need to move here due to dependencies. The only new code is then pulling a new ResolutionControl class out of the existing resolver UI so that stepping through the resolver is not tied to the UI. Although there are several other files touched, they are only due to the package name changes.

By itself this change doesn't do much other than make the resolver code a little cleaner/organized and allow it to be reused elsewhere. However, I still think it's beneficial to clean this up by itself, and I have another change ready that'll add resolution info to the CDS.