Admin button to apply end of updates
With the fix in PR #234 I have realized that there have been a few times in the last couple years where we've run the resolver on a contest that is over, final, and has no unjudged runs, but didn't have the end_of_updates flag set. Prior to the fix this would have gone unnoticed, now (with the fix) it will correctly block the resolver from starting.
This fix just adds an admin button that would apply the end of updates in a pinch, so that you wouldn't need to go back to the CCS or manually edit the event feed if this happens.