CDS resolution support
Basic support for the CDS to participate in the resolving process. Adds a new section to the admin page that is opt-in. If you enable, it shows the status of the resolving, allows you to control the stepping (not normally expected to be used) and exposes resolved judgements to public contest clients (for instance, you'd see the judgements on a presentation scoreboard).