Show freeze on scoreboards
At NAC someone requested that the clock change on the main scoreboard so that you can tell if it's during the contest freeze. Adding additional text makes it too long, so I've just changed the colour. Tried blue but it was too pale so I went with yellow instead. Also added the ability to disable the clock on the tile presentations, since you could do this with the other scoreboards (in case of contest pause or some other unforseen case). Updated the presentation properties for the admin, including a bunch that were missing.