Move .cache files to temp folder
When the disk contest source is being used to load a local contest I moved the cache files to a temp folder. To avoid a large directory tree with only one file per folder I switched to a flat set of files in this case.
As part of this I had to make the cache map and a couple methods not static, but that's a good idea anyway: contests don't overlap on disk, each having their own (smaller) cache is better and there's no reason to hold up loading of one contest due to another one.