Show CCS connection status
One of the issues raised before in issue #55 (and actually happened at NAC) was that we don't highlight the CCS connection status well enough in the web UI: if the CCS feed or the network goes down you might not notice. This commit changes two things: in the homepage instead of "Live" it will now show "Connected to CCS" or "CCS connection error" (in red) for every contest, and in the more detailed overview it'll show the full connection state (initializing, reading, reconnecting, failed, etc).