Add warning indicator to page header
Followed Nicky's suggestion and added an indicator to the page header so that you'll see the ! warning on every page, even if you're not looking at that contest right now.
Since I was adding it in one more place, I turned the error text into a getError() method on ConfiguredContest - for now it only returns the CCS connection error, but it could be used to flag other errors in the future.
Also switched to correctly using bootstrap for the CCS.