Single tools version
We have never versioned the ICPC tools separately, and do not intend to. Even if we were going to have a release for just one tool we'd bump the version, release it, and skip that version for the other tools.
So, I'm moving the version file to the root, removing it from each of the tools, and fixing the build.xmls. This is simpler to do a release, and no chance of missing one or getting out of sync.