Form-based login
The CDS was using BASIC auth for everything. This had a few drawbacks, including no fancy login page, and no way to logout (and hence no way to switch users either). This change adds FORM auth for the main website (which works if you cross over to API) and keeps BASIC auth for API & presentation websocket