Switch link parser to better XML parser
The link parser (used to determine if there are newer copies of any of the clients available on the CDS) is broken by the AdminLte upgrade. Last time this happened it was the same issue (&nbsp is not understood by the SAX parser) and I just removed the space, this time I'll switch to a parser that can handle it.