DisplayConfig - start cleaning up multi-screen support and improve admin data
Prior to this change, the only way to access multi-screen support was an undocumented admin property, e.g. 'pos:220'. (this tells the client that they are position 0 [top left] in a 2x2 grid of monitors) There was no way to launch clients like this, use a second monitor in this way, nor could the admin tell afterward that any clients were in this mode (they only report being in full screen or not). Likewise there is an undocumented client mode for testing where you can use '--display 2a' to put a client at the top left of the second display.

To start cleaning this up, I've merged these two things into a new DisplayConfig class. Use of this is still undocumented, but there are two benefits:
 - Clients can use parameters like '--display 2322' (display on second monitor, we are position 2 [top-right] in a 3x2 grid) or '--display 1c222'(use the bottom-left of the monitor as the bottom-left of a 2x2 grid).
 - Clients send the display config to the server, so the admin is now aware of any of these modes. It doesn't display anything different yet, but I started to add multi-selection awareness.

Notes:
 - I left a to-do, because I intend to tackle that next...
 - If you can think of a better way to represent these modes in a string, please feel free to suggest. :)
 - The FPS wasn't always shown in a consistent font, fixed that.