Command line support for --multi-display
Clean up and expose the --multi-display option so that it can be used from the command line. Added basic validation.
Added to presentation clients & resolver. Although the last is unlikely to be used, you never know and it keeps all clients in sync.
Did not updated the readmes since they already do not show all cli options. IMHO this is ok, advanced options can just be shown in --help.