Reorder log file setup and option parsing
The resolver logs to one of three log files to make it easy to have client and server running on the same machine and view the logs - but b/c option parsing happens before this you would see output like "Option found: --speed 0.3" on the command line. Reordering where this happens to avoid log file escapes.