Support submission interface
First commit to support POST for submissions. RESTContestSource is updated to provide a method to perform a submit against a Contest API that supports it and the CDS' ContestRESTService supports POST by team or admin. The CDS acts as a simple proxy to the CCS.

Some details may need to be tweaked based on the final spec, but this has been tested with a local client able to submit to either the CDS or a temporary DOMjudge instance.