Improve submission interface
Changed the submit() method to allow an optional files parameter that does the work of encoding for you. This makes it easy to do proxying or a simple client pattern where you just specify the attributes and file(s):
  JSONObject obj = new JSONObject();
  obj.put("language_id"...);
  submit(obj, file);

Tracing is logged to info (log file) rather than error, which outputs to the console - whoever calls these methods should determine how to tell the user it didn't succeed. I also created a new JSONWriter that makes it easy to output a JSONObject to a stream. There's a lot of overlap with JSONEncoder, but it didn't make sense to merge. I may go back to users of that class to see if the writer works better.