Handle Windows scaling
When I was working on the last PR I noticed that the presentation window wasn't covering the full screen on a display that Windows was scaling. After a bit of digging and trial and error I've found how to correctly scale the window.

At the same time I fixed a separate bug: the display config to use the max screen resolution (e.g. "--display 1x") wasn't working on Windows because it requires you to enter full screen exclusive mode before setting the display mode. I switched the order of those.

I also added calls to (re-)create the background buffer strategy after changing the display - this is most likely overkill, but felt safer.

One big caveat: Windows scaling can only be used on the primary display. Based on testing, stack overflow, and googling, there is no way to support a secondary display using scaling. I can get the window position correct, but then can't draw on it. This is likely a bug that could be fixed in a future JDK, but in the meantime the scaling needs to be 100% on all non-primary displays.