Team display style improvements
Currently the ability to override the default team naming behaviour (usually the team's display name) is only visible as an option to the resolver. Due to some issues at a regional we want to enable this in the other tools and give a more wide open way to specify the name.

- Change style to a string.
- Change from enum of options like team_name to string substitution like "{team.display_name} - {org.name}". {team.name} and {org.formal_name} are also supported.
- Give a better way of setting the default (vs setting on every presentation) and use it in the resolver.
- Enable --style option in both presentation clients (matching what the resolver had).
- Fix up all the dependent presentations.

Readme updates will follow...