Second backup for team award presentation
A few years ago we changed the team award presentation from using the ICPC logo to the contest logo from the REST feed. However, John has had a couple contests now that don't have a logo and the screen is blank, so let's add a second backup and bring back the ICPC logo if there is no contest logo.