Improve performance when deleting many contest objects
The contest model (ContestData) is optimized for fast lookup and additions. While debugging a problem for John I found that in some edge cases (e.g. large contest, with run info, filtering down to even half the teams) the performance was awful, adding a full minute to the resolver startup time.
I finally bit the bullet and added a proper bulk delete operation. There is probably still a better way to do it or optimize, but this cuts the same 60s operation down to 6-7s.