Update presentation list in the readme
In the past we've talked about a live list of the presentations being generated into the docs during the build, but that's a big pain. As a stopgap/step towards that, I just changed the output of the command line to markdown, which I could then just copy/paste into the readme. It's still manual, but at least it's an easy step. I was careful to keep the command output reasonably clean so that it's still as usable as possible directly from the command line.
While I was at it, I improved several presentation descriptions or made them match what was in the docs previously.