Add Award dialog
Added a button that opens a dialog (dropdown for award type, citation, etc) to add any award to the contest. Awards can be added to multiple teams at once, and there is a warning if the award you're giving has already been assigned and will be reassigned (maybe should also be a secondary confirmation in the future?).
While I was doing that I also fixed a bug assigning medals in the AwardUtil, a minor formatting problem in the Edit dialog, and improved the award template dialog quite a bit - you can now type a template, paste, or load from file and it'll read the template in NDJSON or JSON array formats.