Determine base URL of API correctly.
Instead of assuming the API of external CCS'es always starts with /api/, determine the real URL by using a regex on the CCS URL.
Kattis uses /clics-api/contests/xxxxx for a contest so it broke there if they use relative URL's.
Also use the root of the URL instead of the part before /api/ for absolute URL's, since I think that is what is meant in the spec.