First pass at team (non-admin) web UI
I created a team version of the UI that doesn't have API http links and 'admin-y' things like element ids. I don't _love_ this change, but I couldn't come up with a cleaner way to do this without creating copies of several files - although the UIs are close, I think you'd end up with a lot of if statements and odd logic to handle the number of columns and other differences.

Here's what I did:
- Removed the Overview page for teams in addition to the others. This is really just a 'connection details' page, so maybe it should be renamed even for admins.
- Removed the 'contest.setContestId()' from as many pages as possible and renamed them from .jsp to .html to be clear that they're now 'clean' client-side code.
- Renamed most existing /details pages to -admin and created a second copy of each for teams that removes details like API links, object ids, and some details that I don't think teams would ever need.
- Changed the layout a bit and removed the ability to collapse many of the sections.
- Removed awards, submissions, and judgements sections from the Details for teams.

I would like to have a better UI for some things (e.g. clarifications should probably be on their own page in a post-reply kind of UI vs the current table) but this is a pretty good pass at making it generally ok. There are a couple things that still aren't 'clean': the submissions page contains ids and the scoreboard page has a link to the API. I didn't touch either of these yet because both are minor, and both pages also contain server-side logic (.jsp) that ideally I'd change to client side (javascript) first.