contest.js class
Change the contest.js to a proper class. This will allow some pages like the welcome page to connect to multiple contests at the same time and creates a better interface for clients. I also started using => to simplify the code.
This means we no longer support IE and currently 5% of the worldwide browsers according to https://caniuse.com/?search=class. This is unfortunate, but this is the cleanest and correct way to clean up the contest support for future and that number should be dropping.