Initial Commentary presentation
This is an initial commentary presentation as requested by Fredrik. This pass is functional and similar to a judge queue - commentary slides up from the bottom when they occurs, then fade away after a while - but is very vanilla. The messages are shown as-is, no formatting of teams or problems, text is not wrapped, etc. I'll follow up with visual improvements later.