Multiple file download
Fixes a problem where file references were set to null on objects that had multiple downloads. Cleans up the support in general by looking at local disk and syncing all files before trying to attach anything to the local contest object.
Also updates country_flag to the location that we went with in the CAF proposal.