Contest model support for SVGs as images
Adds basic SVG support to the Java contest model. When a client requests an image of a specific size SVGs are used identically to pngs or jpgs. The code finds the 'best' image the same way it as always has based on the size requested and what's available, then will create a custom-sized BufferedImage using the svg if that's the best or only option available.
There are probably going to be some rare cases in the future where it'd be better for a particular client to natively understand the differences of an SVG and paint directly, but those are actually going to be rare - most presentations are creating composite images (e.g. a row with team logo and name) and this is all they need. This change brings all the benefits of svgs (small download, better resizing) to these clients in a completely transparent way.