Local feed + REST contest source
There are currently 3 ways to load a contest (6 if you count XML vs JSON...): 1) HTTP URL/REST (with absolute or relative file references), 2) Disk/CDP/CAF (with local resources), and 3) directly from feed file (with no references, since this is mostly historic).
This switches the third method to use the REST contest source from the first, which means that if there are absolute file references in the feed then it should be able to download and use those. Relative URLs still won't work, since the behaviour is undefined and wouldn't match the CAF.
Also using this commit to slip in the one line change in DiskContestSource that was missing from previous commit.