Coach view support for multiple streams
Coach view support for multiple desktop and webcam streams per team. If there are multiple streams for a team they are shown as two entries in the list, e.g. "UBC (1)" and "UBC (2)". Selecting either will give the first pair of webcam/desktop or second. If the number of webcams and desktops don't match, the maximum number of options are shown and you'll only see the one stream that exists for those options.