Award Util support for PUTing awards
Added a button to the Award UI named "Upload" that is only enabled when you're connected to a contest via REST. Clicking this button PUTs all awards to the server. This overwrites any of these award ids that exist (provided you have access) but does not affect any other award ids. As such there's a warning when you click the button and we need to decide what the options/behaviour should be long term.

To enable this I found it was more convenient to have post/patch/put/delete methods that accept a contest type + object instead of a json object. I also made the methods generic so they can accept any contest object. This means the methods can't have type-specific information, but that's covered in the spec and means we don't need separate methods for every type.

Also noticed that the top award generation was making the award id "top-top-x", so fixed that.