Fix missing app icons
Application icons (e.g. on the dock on Mac) were missing for SWT apps. Not sure if it's a regression or how, but fixed this by using the same code that was being used for AWT apps. Made a common method in ContestModel - even though it's not the 'correct' place, it's one place where all apps can use it from.