Support for Honors (score compared with a team at a given percentile) awards
Added support for awards based on a range of teams that have a score less than a given percentile. This allows us to do:
`{"id":"honors-high","parameter":"0-25", "citation":"High Honors"}
or
`{"id":"honors-regular","parameter":"25-50", "citation":"Honors"}`
or
`{"id":"honors-mention","parameter":"50-100"}`
(citation automatically defaults to Honorable Mention)

For the record, it was seriously difficult to stick to US English here and not spell honourable correctly. ;-)