Support for reading config from 2020 CMS
Am I happy with this code? No. Is it likely to require changes even in the next week? Yes. Cleanup afterwards? Are there some hardcoded bits? Absolutely.

The bulk of the change is reading directly into Contest API objects instead of the intermediate CMS objects that were there before. This means that you can import from .tsv or other formats, do any manipulation using standard mechanisms, and easily save to .tsv or .json formats. I had this part in a nice simpler commit but messed up my git history so it'll have to go in as one chunk now.