Team member machine-specific API
As per the last CLICS call, I am checking in the API to allow team-member video, desktop, backups, etc, for cases where each team member has their own machine, and it is known which team member is using which machine. We do not know how or if we'll be able to do the mapping for the 2020 World Finals (or if this is a one-off), but this provides the API in order to allow it.