Move resolver UI code where it belongs
The unchanged 'TODO' in ResolverLogic 689 was bugging me, so I went back and fixed it properly - how to adjust the scrolling to account for how many teams are shown onscreen just shouldn't be in the generic resolver logic. I moved this to the UI side where it can correctly adjust for any changes.