Start to standardize --bill
--bill was created a few years ago, and it amounted to two things:
1. Use judge queue for Honorable Mention teams.
2. Show a list for Honorable Mention teams and subsequent groups of teams that solve each consecutive # of problems (except medals).

We've talked about how to make it a bit more standard, and this is the first step of that, for #2. Up to now the award generator/resolver has used a boolean attribute on awards called "show" that was either false (just pause, then move on, e.g. for FTS) or true (stop and show the team detail, e.g. for medals). This change turns it into a string display_mode attribute that has three options: "detail", "pause" (the same two options as before) and "list". Any award that has this last option will wait until every team has been resolved and then show the full list of teams. The code is backward compatible with old feeds and for now --bill creates hidden awards to stop at the same places as before.

Why all this? Besides cleaning up this custom code it means you can have awards for things like Honors and decide to show a list.