Improve text fit when names are too long
It was annoying me that the team intro presentation logos weren't always near the text and sometimes too wide for the screen. Improved the text helper to correctly handle alignment and scaling to fit within a specific bounds. Also improves text fitting on the scoreboard and tile scoreboard when names are too long.