2020 CMS utils
This is extremely ugly, but these are the last changes that I got working ~1 week before Moscow World Finals. At least some aspects were broken by the time we got onsite, so use at your own risk... checking in for posterity and so we have the full history to start on next year.