Add CDS name to web
If you name a CDS (either through non-default folder name, or adding <cds name="My name"> to the cdsConfig.xml), we'll replace the standard CDS title at the top left and the page title so that you can tell multiple CDSs apart. Also updated the About page, and reordered the welcome include b/c it was affecting the header and footer.