Fix and update floor map
Fix regressions with floor map that Nicky reported, fix issues writing the extended MapInfo to disk, and break backward compatibility (figured it's a good time due to the floor map switch to json) to make problem location consistent with team.

- Team: remove support for old group_id, and standalone x/y/rotation attributes (they've been inside a location object for a few years). Add setLocation() helper.
- Problem: move x/y into a location object to match team, for addImpl() use switch statement instead of lots of ifs, make test data count optional again. Add setLocation() helper.
- ContestWriter: add support for writing singleton objects like State and MapInfo. (missed that before)
- MapInfo: change printer helper to be called setLocation to match others, and use the setLocation() helpers.
- FloorMap: use setLocation() helpers, clean up rotating entire floor.
- FloorGenerators: load existing contest from disk so that the problem ids match.

There is still a small regression in that the floor map UI shows only the data about to be written to the floor map - there are no problem labels (since we don't really want them duplicated in the floor map data) so it uses the longer problem ids. There are a couple options here (e.g. copy labels from one contest to the other, or merge before showing), but I haven't decided which is best (or if it's actually necessary) so doing neither for now.